OMS UserController - Code Review fix point 2
