+++ OMS.API/Web.config
@@ -22,7 +22,7 @@
       <logger name="*" minlevel="Info" writeTo="logfile" />
     </rules>
   </nlog>
-  <appSettings configSource="AppConfig.config"/>
+  <appSettings configSource="AppConfig.config" />
   <system.web>
     <compilation debug="true" targetFramework="4.6.1" />
     <httpRuntime targetFramework="4.6.1" />
@@ -120,5 +120,5 @@
       <provider invariantName="System.Data.SqlClient" type="System.Data.Entity.SqlServer.SqlProviderServices, EntityFramework.SqlServer" />
     </providers>
   </entityFramework>
-  <connectionStrings configSource="ConnString.config"/>
+  <connectionStrings configSource="ConnString.config" />
 </configuration>
