+++ OMS.BusinessGateway/Task/User.cs
@@ -1,19 +1,39 @@
-﻿using OMS.DataGateway.Repositories;
+﻿using Newtonsoft.Json;
+using OMS.DataGateway.Repositories;
 using OMS.DataGateway.Repositories.Iterfaces;
 using OMS.DomainGateway.Task;
 using OMS.DomainObjects.Request;
 using OMS.DomainObjects.Response;
-using System;
+using RestSharp;
 using System.Collections.Generic;
+using System.Configuration;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Domain = OMS.DomainObjects.Objects;
 
 namespace OMS.BusinessGateway.Task
 {
     public partial class BusinessUserTask : UserTask
     {
         private readonly IUser _userRepository;
+
+        #region Private Methods
+        private static string GetApiResponse(string apiRoute, Method method, object requestQueryParameter, string token)
+        {
+            var client = new RestClient(ConfigurationManager.AppSettings["ApiGatewayBaseURL"]);
+            client.AddDefaultHeader("Content-Type", "application/json");
+            if (token != null)
+                client.AddDefaultHeader("Token", token);
+            var request = new RestRequest(apiRoute, method) { RequestFormat = DataFormat.Json };
+            request.Timeout = 500000;
+            if (requestQueryParameter != null)
+            {
+                request.AddJsonBody(requestQueryParameter);
+            }
+            var result = client.Execute(request);
+            return result.Content;
+        }
+        #endregion
+
         public BusinessUserTask(IUser userRepository)
         {
             _userRepository = userRepository;
@@ -33,12 +53,78 @@ namespace OMS.BusinessGateway.Task
 
         public override UserResponse CreateUpdateUser(UserRequest user)
         {
-            //If needed write business logic here for request.
+            UserResponse userResponse = new UserResponse();
 
-            UserResponse userData = _userRepository.CreateUpdateUser(user);
+            foreach (Domain.User userDetails in user.Requests)
+            {
+                #region Create TMSUserRequest
+                UserRequest tmsRequest = null;
+                if (userDetails.Applications.Contains(2))
+                {
+                    tmsRequest = new UserRequest()
+                    {
+                        Requests = new List<Domain.User>()
+                        {
+                            new Domain.User()
+                            {
+                                Applications = new List<int>(){
+                                    2
+                                },
+                                FirstName = userDetails.FirstName,
+                                LastName = userDetails.LastName,
+                                UserName = userDetails.UserName,
+                                Password = userDetails.Password,
+                                ConfirmPassword = userDetails.ConfirmPassword,
+                                Email = userDetails.Email,
+                                PhoneNumber = userDetails.PhoneNumber
+                            }
+                        },
+                        CreatedBy = "SYSTEM",
+                        LastModifiedBy = user.LastModifiedBy,
+                        CreatedTime = user.CreatedTime,
+                        LastModifiedTime = user.LastModifiedTime
+                    };
+                }
+                #endregion
 
-            //If needed write business logic here for response.
-            return userData;
+                #region CreateUpdateUser in OMS
+                if (userDetails.Applications.Contains(1))
+                {
+                    userResponse = _userRepository.CreateUpdateUser(user);
+                }
+                #endregion
+
+                #region CreateUpdateUser in TMS
+                if (tmsRequest != null) //For TMS Application - Integrate Azure API Gateway
+                {
+                    LoginRequest loginRequest = new LoginRequest();
+                    //Login to TMS and get Token
+                    string token = string.Empty;
+                    loginRequest.UserName = ConfigurationManager.AppSettings["TMSLogin"];
+                    loginRequest.UserPassword = ConfigurationManager.AppSettings["TMSPassword"];
+                    var tmsLoginResponse = JsonConvert.DeserializeObject<UserResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayTMSURL"]
+                        + "/v1/user/login", Method.POST, loginRequest, null));
+                    if (tmsLoginResponse != null && tmsLoginResponse.Data.Count > 0)
+                    {
+                        token = tmsLoginResponse.TokenKey;
+                    }
+
+                    UserResponse tmsUserResponse = JsonConvert.DeserializeObject<UserResponse>(GetApiResponse(ConfigurationManager.AppSettings["ApiGatewayTMSURL"]
+                         + "/v1/user/createupdateuser", Method.POST, tmsRequest, token));
+
+                    if (!userDetails.Applications.Contains(1))
+                    {
+                        userResponse = tmsUserResponse;
+                    }
+                    else
+                    {
+                        userResponse.StatusMessage = userResponse.StatusMessage + ". " + tmsUserResponse.StatusMessage;
+                    }
+                }
+                #endregion
+            }
+
+            return userResponse;
         }
 
         public override UserResponse DeleteUser(int UserID)
